Refactor preprocessingsnippets
Rename it to processingstandards
Shift code around
Improve readability of the snippets
